pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-3d-programming-book"
version = "1.0.0"
description = "A comprehensive guide to Python programming and 3D graphics development"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Python & 3D Programming Book Contributors", email = "your-email@example.com"}
]
maintainers = [
    {name = "Python & 3D Programming Book Contributors", email = "your-email@example.com"}
]
keywords = ["python", "3d", "graphics", "opengl", "education", "tutorial", "book"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Education",
    "Topic :: Multimedia :: Graphics :: 3D Rendering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Visualization",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "PyOpenGL>=3.1.0",
    "PyOpenGL-accelerate>=3.1.0",
    "Pillow>=8.3.0",
    "matplotlib>=3.4.0",
    "glfw>=2.0.0",
    "scipy>=1.7.0",
    "sympy>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.12.0",
    "flake8>=3.9.0",
    "black>=21.0.0",
    "mypy>=0.910",
    "pre-commit>=2.15.0",
    "bandit>=1.6.0",
    "safety>=1.10.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=0.5.0",
    "myst-parser>=0.15.0",
]
examples = [
    "jupyter>=1.0.0",
    "ipywidgets>=7.6.0",
    "plotly>=5.0.0",
    "bokeh>=2.3.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/python-3d-programming-book"
Documentation = "https://github.com/yourusername/python-3d-programming-book#readme"
Repository = "https://github.com/yourusername/python-3d-programming-book.git"
"Bug Tracker" = "https://github.com/yourusername/python-3d-programming-book/issues"
"Discussions" = "https://github.com/yourusername/python-3d-programming-book/discussions"

[project.scripts]
python-3d-book = "python_3d_programming_book.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.py"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["python_3d_programming_book"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["python_3d_programming_book"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "docs", "examples"]
skips = ["B101", "B601"]

[tool.safety]
output = "json"
